+++ st2reactor/st2reactor/cmd/sensormanager.py
@@ -45,14 +45,15 @@ def main():
     try:
         _setup()
         container_manager = SensorContainerManager()
-        sensors = _get_all_sensors()
+        sensors = None
 
-        if cfg.CONF.sensor_name:
-            # Only run a single sensor
-            sensors = [sensor for sensor in sensors if
-                       sensor.name == cfg.CONF.sensor_name]
-            if not sensors:
-                raise SensorNotFoundException('Sensor %s not found in db.' % cfg.CONF.sensor_name)
+        if cfg.CONF.sensor_ref:
+            sensor = SensorType.get_by_ref(cfg.CONF.sensor_ref)
+            if not sensor:
+                raise SensorNotFoundException('Sensor %s not found in db.' % cfg.CONF.sensor_ref)
+            sensors = [sensor]
+        else:
+            sensors = _get_all_sensors()
 
         if not sensors:
             msg = 'No sensors configured to run. See http://docs.stackstorm.com/sensors.html. ' + \