+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ in development
   file over if it exists. (bug fix)
 * Support for masking secret parameters in the API responses. Secret parameters can only be viewed
   through the API by admin users. (new-feature)
+* Single sensor mode of Sensor Container uses ``--sensor-ref`` instead of ``--sensor-name``.
 
 0.11.3 - June 16, 2015
 ----------------------